+++ src/Core/Sql/Schema.sql
@@ -183,6 +183,8 @@ INSERT INTO Configuration (Name, Value) VALUES('PreviewAutoPlayVideo', 'false');
 INSERT INTO Configuration (Name, Value) VALUES('PreviewAudioMeters', 'true');
 INSERT INTO Configuration (Name, Value) VALUES('PreviewTemplates', 'true');
 INSERT INTO Configuration (Name, Value) VALUES('WarnMissingMedia', 'true');
+INSERT INTO Configuration (Name, Value) VALUES('RelayCheckInUrl', '');
+INSERT INTO Configuration (Name, Value) VALUES('RelayCheckInToken', '');
 INSERT INTO Configuration (Name, Value) VALUES('ShowSTEPButton', 'true');
 INSERT INTO Configuration (Name, Value) VALUES('HotkeyToggleAutostep', 'Ctrl+Shift+N');
 INSERT INTO Configuration (Name, Value) VALUES('HotkeyToggleAutostepAlt', '');